+++ semgrep-core/tests/scala/dots_args.scala
@@ -1,0 +1,12 @@
+object Foo {
+
+def foo() {
+  //ERROR:
+  foo(1,2,3,4,5)
+  //ERROR:
+  foo(5)
+}
+}
+
+
+