+++ semgrep-core/tests/scala/dots_string.scala
@@ -1,0 +1,10 @@
+object Foo {
+def foo() {
+    //ERROR:
+    foo("whatever sequence of chars")
+    //ERROR:
+    foo(s"whatever sequence of chars")
+    //ERROR:
+    foo("""whatever sequence of chars""")
+}
+}