+++ semgrep/semgrep/semgrep_types.py
@@ -44,6 +44,7 @@ class Language(Enum):
     KOTLIN: str = "kt"
     YAML: str = "yaml"
     ML: str = "ml"
+    SCALA: str = "scala"
     JSON: str = "json"
     REGEX: str = "regex"
     GENERIC: str = "generic"
@@ -66,6 +67,7 @@ class Language_util:
         Language.RUST: [Language.RUST.value, "Rust", "rs"],
         Language.KOTLIN: [Language.KOTLIN.value, "Kotlin", "kotlin"],
         Language.YAML: [Language.YAML.value, "Yaml"],
+        Language.SCALA: [Language.SCALA.value],
         Language.ML: [Language.ML.value, "ocaml"],
         Language.JSON: [Language.JSON.value, "JSON", "Json"],
         Language.REGEX: [Language.REGEX.value, "none"],